feat(dashboard): make v2 the default at /dashboard, move legacy to /dashboard/legacy
Done at the user's request — the v2 console has been the staging surface
since #59 and the recent PRs (#60–#66) closed the data hygiene, mobile,
ratings UI, deep work, T3T autosave, and money-note gaps. Time to flip it.

# Route topology

- /dashboard         → v2 (was at /dashboard/v2). Source:
                       src/app/dashboard/page.tsx.
- /dashboard/legacy  → old dashboard (was at /dashboard). Source:
                       src/app/dashboard/legacy/page.tsx.
- /dashboard/v2      → permanent server-side redirect to /dashboard so
                       old bookmarks / external links resolve. Single-file
                       server component using next/navigation::redirect().

# Hand-off details

- Impersonation (/as/[role]/dashboard) still renders the LEGACY dashboard
  because the impersonation banner was visually validated against it.
  Migrating impersonation to v2 needs the banner re-checked against the
  dark shell — tracked in the arch doc.
- DashboardPage jest suite still imports the legacy page since its
  assertions are tied to the legacy layout. The v2 surface has its own
  test files under src/components/dashboard/v2/__tests__.
- The "← Old dashboard" link in the v2 header was renamed to "Legacy"
  and now points at /dashboard/legacy. data-testid: legacy-dashboard-link.

# Tests / specs

- All 18 Playwright cases in dashboard-v2.spec.ts and 13 in
  t3t-autosave.spec.ts now navigate to /dashboard instead of
  /dashboard/v2. Spec headers updated. The parity test renamed to
  "parity: /dashboard shows the same temporal hours as /api/focus-hours"
  to reflect the new route.
- 42 jest suites, 833 tests still pass.
- next build verified all 3 dashboard routes register as ○ static:
  /dashboard, /dashboard/legacy, /dashboard/v2 (the redirect).

docs/dashboard-v2-architecture.md updated with the new route topology,
the future-cleanup checklist (migrate impersonation, delete v2 redirect
file once external refs are gone), and updated ASCII diagram.
